ADD comments for UserResponse model and its builder
